+++ src/layouts/DashboardHeader/types.ts
@@ -2,7 +2,7 @@ import { ReactNode } from "react";
 
 import { MenuItemType } from "lib";
 
-import type { NavbarPropsType } from "components/Navbar/types";
+import type { NavbarPropsType } from "components/app/Navbar/types";
 
 export type DashboardHeaderPropsType<MenuKeys> = {
   menuMap: MenuItemType<MenuKeys>[];